bnd-maven-plugin: Avoid writing output if not out-of-date
The Builder collects that latest modified time of it source material
such as bnd files and included files. The Jar collects the latest
modified time of all its contents including the Builder which built it.

So we modify the expandJar method to only write files out when either
the
file does not exist or the file is older than the Jar last modified
time.
For incremental builds, this should avoid writing any output files if
none of the input material to the built Jar has changed. Thus we avoid
downstream dependencies from rebuilding because unchanged files were
rewritten.

Alternative to https://github.com/bndtools/bnd/pull/1316

Signed-off-by: BJ Hargrave <bj@bjhargrave.com>