bnd-maven-plugin: Add 'fastIncrementalCheck' property
By defining this property in your POM (or any parent POM) a fast
incremental check is supported. Before any bnd processing is started on
incremental build the Plexus Build API is used to determine if the
project has an changes. If it does not, nothing will be done. 

If however this property is enabled and on a full build source meterial
is discovered that lies outside the project directory, the build will
fail with an error message. In this case you need to disable or remove
'fastIncrementalCheck'